// frontend/src/security/types.ts
export interface IAvailability {
  /**
   * Some Menu Items will only appear when the logged in Company is from a specific
   * country. For example Australia has some features that are unique and therefore
   * shouldn't appear for the UK. The string must be a two letter ISO country code
   * to avoid type mismatches with the GraphQL API.
   */
  country?: [string];
  /** Different Security roles will have access to different functionality. Restricting
   * the Menu Items by Security roles, means security becomes a configurable issue,
   * rather than a code issue.
   */
  security?: Array<SecurityRole>;
  /** Currently we can restrict menu items to Companies that are of a certain Business
   * Type. Currently only "HEALTH" is a supported option.
   */
  businessType?: Array<string>;
}

/**
 * This is just combination of IAvailability & LicensePlan
 * This will be used on IRoute, IMenuItem and ITab
 */
export interface IContentSecurity {
  /**
   * Conditions to determine whether this Item is available to the User and
   * the Company.
   */
  availability?: Array<IAvailability>;
  /**
   * Conditions to determine whether user needs to upgrade his plan or not
   * Some tabs/menu/routes will be restricted by the License that the company has. For example, a
   * Company may be in the FREE plan, but the Tab is only available on the PRO or
   * ENTERPRISE license. In that instance the Tab should still appear, but a message
   * will be displayed to the User to upgrade their Company plan
   */
  licenseRequired?: LicensePlan[];
}

/**
 * The Menu item appear in the Drawer of the application and allow the User
 * to navigate between sections.
 */
export type IMenuItem = IContentSecurity & {
  name: string;
  /**
   * The order the Menu Item appears in the Drawer
   */
  order: number;
  /**
   * Set Icon
   */
  icon?: () => React.ReactNode;
  /**
   * set url to navigate
   */
  url: string;
  /**
   * set logic to determine whether menu item should be shown as active or not
   * function should return true/false
   */
  isActive?: (pathname: string) => boolean;
};
// ...
export enum SecurityRole {
  admin = 'ADMIN',
  user = 'USER',
  root = 'ROOT',
  owner = 'OWNER',
  healthcareProvider = 'HEALTHCARE PROVIDER',
}

export enum LicensePlan {
  PRO = 'PRO',
  ENTERPRISE = 'ENTERPRISE',
  FREE = 'FREE',
}